# api/openapi.yaml
openapi: 3.1.0
info:
  title: CloudOptix API
  version: "1.0.0"
  summary: >-
    Autonomous cloud cost intelligence and optimization: architecture-aware
    discovery, economics, recommendations, governed automation and an
    auditable execution trail, all tenant-isolated.
  description: |
    CloudOptix models a customer's AWS estate as a digital twin, prices it,
    finds savings, and — within policy — executes and rolls back changes
    itself. This document describes the whole HTTP API a client (the
    CloudOptix web app, a CI pipeline calling the cost compiler, or a partner
    integration) talks to.

    ## Authentication
    Every route except `/healthz`, `/readyz`, `/metrics`, `/openapi.yaml` and
    the `Onboarding` tag requires a credential: either an OIDC-issued bearer
    JWT (`Authorization: Bearer <token>`) or a service API key
    (`X-CloudOptix-Api-Key: <key>`). See the `BearerAuth` and `ApiKeyAuth`
    security schemes below. A worker or CI system uses a service token
    (issued out of band) presented the same way as a user's bearer token.

    ## Tenant scope
    An OIDC principal may belong to more than one tenant; which one a
    request is scoped to is resolved from the token's own tenant claim when
    present, or from the `X-CloudOptix-Tenant` header otherwise, with a
    membership check either way — a header can select among tenants a
    principal actually belongs to, never grant a new one. A service token
...
    ever names a tenant explicitly.

...

...

...

...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# YAML anchors for parameters and error responses shared across dozens of
# operations. A vendor extension at the document root (x-*, permitted
# anywhere per the OpenAPI 3.1 Specification Extensions rule) rather than a
# components/ entry, because these are resolved by the YAML parser itself —
# before any OpenAPI tooling sees the document — while a $ref into
# components/parameters would still repeat one line per operation for no
# benefit; the effect on the parsed document is identical to writing every
# field out longhand at every one of the ~115 operations below, just without
# ~115 chances for one of those copies to drift from the rest.
...
paths:
...
  /aws-accounts:
    parameters: [*tenantHeader]
    get:
      operationId: aws_accounts.list
      tags: [AWS Accounts]
      summary: List connected AWS accounts
      responses:
        "200":
          description: All accounts.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AWSAccount" }
        <<: *commonErrors
    post:
      operationId: aws_accounts.register
      tags: [AWS Accounts]
      summary: Register a new AWS account
      parameters: [*idempotencyKey]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterAccountRequest" }
      responses:
        "201":
          description: Registered, with connection instructions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  account: { $ref: "#/components/schemas/AWSAccount" }
                  next_steps: { $ref: "#/components/schemas/AWSOnboardingInstructions" }
        <<: *commonErrors
  /aws-accounts/{id}:
    parameters: [*tenantHeader, *idPathParam]
    get:
      operationId: aws_accounts.get
      tags: [AWS Accounts]
      summary: Get one AWS account
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AWSAccount" }
        <<: *commonErrors
    delete:
      operationId: aws_accounts.remove
      tags: [AWS Accounts]
      summary: Disconnect an AWS account
      parameters: [*idempotencyKey]
      responses:
        "204": { description: Removed. }
        <<: *commonErrors
...
  /aws-accounts/{id}/suspend:
    parameters: [*tenantHeader, *idPathParam, *idempotencyKey]
    post:
      operationId: aws_accounts.suspend
      tags: [AWS Accounts]
      summary: Suspend an AWS account (stop discovery and automation)
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SuspendAccountRequest" }
      responses:
        "204": { description: Suspended. }
        <<: *commonErrors
  /aws-accounts/{id}/instructions:
    parameters: [*tenantHeader, *idPathParam]
    get:
      operationId: aws_accounts.instructions
      tags: [AWS Accounts]
      summary: Get connection instructions again
      responses:
        "200":
          description: Instructions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AWSOnboardingInstructions" }
        <<: *commonErrors

  # ================================================================
  # Discovery
  # ================================================================
...
  /costs/summary:
    parameters: [*tenantHeader, *fromParam, *toParam]
    get:
      operationId: costs.summary
      tags: [Costs]
      summary: Get the cost-intelligence headline
      responses:
        "200":
          description: Summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CostSummary" }
        <<: *commonErrors
...
  /recommendations/summary:
    parameters: [*tenantHeader]
    get:
      operationId: recommendations.summary
      tags: [Recommendations]
      summary: Get the recommendations headline
      responses:
        "200":
          description: Summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecommendationSummary" }
        <<: *commonErrors
...
  /approvals:
    parameters: [*tenantHeader, *cursorParam, *limitParam, *sortParam, *orderParam]
    get:
      operationId: approvals.list
      tags: [Approvals]
      summary: List approval requests
      responses:
        "200":
          description: A page of approvals.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalPage" }
        <<: *commonErrors
    post:
      operationId: approvals.request
      tags: [Approvals]
      summary: Manually request an approval
      parameters: [*idempotencyKey]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalRequest" }
      responses:
        "201":
          description: Requested.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalRequest" }
        <<: *commonErrors
...
  /automation/process:
    parameters: [*tenantHeader, *idempotencyKey]
    post:
      operationId: automation.process
      tags: [Automation]
      summary: Run one autonomous optimization cycle
      description: >-
        Finds recommendations the active policy permits to auto-execute,
        plans them, executes them within the maintenance window and
        concurrency limits, and schedules validation. Normally invoked by a
        scheduled worker, exposed here for on-demand and support use.
      responses:
        "200":
          description: Cycle result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutonomousRunResult" }
        <<: *commonErrors
...
  /copilot/suggestions:
    parameters: [*tenantHeader]
    get:
      operationId: copilot.suggestions
      tags: [AI Copilot]
      summary: Get suggested questions for the current tenant state
      description: What keeps an empty copilot screen useful.
      responses:
        "200":
          description: Suggested questions.
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        <<: *commonErrors
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        An OIDC-issued user JWT, an internal service token, or (in
        non-production deployments only) the fixed development static
        token. Validated per internal/infrastructure/auth: issuer, audience,
        expiry and signing algorithm are all checked, and `alg: none` is
        always rejected.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-CloudOptix-Api-Key
      description: A long-lived service API key, bound to exactly one tenant at issuance.
...
  schemas:
...
    RiskLevel:
      type: string
      enum: [NONE, LOW, MEDIUM, HIGH, CRITICAL]
...
    Confidence:
      type: number
      minimum: 0
      maximum: 1
      description: A 0..1 confidence score.
    Money:
      type: object
      description: An exact monetary amount, stored as integer micros to avoid floating-point drift.
      required: [micros, currency, amount, display]
      properties:
        micros: { type: integer, format: int64, example: 4110000000, description: "Exact amount in 1/1,000,000ths of a unit." }
        currency: { type: string, example: USD }
        amount: { type: number, example: 4110.00, description: "Convenience float; micros is authoritative." }
        display: { type: string, example: "$4,110.00" }
    Period:
      type: object
      required: [start, end]
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
...
    AWSOnboardingInstructions:
      type: object
      properties:
        external_id: { type: string, example: "ext-8f2c1e40" }
        trusted_principal_arn: { type: string, example: "arn:aws:iam::111111111111:role/cloudoptix-platform" }
        role_names: { type: object, additionalProperties: { type: string } }
        policy_documents: { type: object, additionalProperties: { type: string } }
        cloudformation_url: { type: string, format: uri }
        terraform_module: { type: string }
        instructions: { type: array, items: { type: string } }
...
    PageEnvelope:
      type: object
      description: The cursor-pagination envelope every list endpoint shares.
      properties:
        next_cursor: { type: string }
        total: { type: integer }

    # -------------------------------------------------------------
    # AWS Accounts
    # -------------------------------------------------------------
    AWSAccount:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        account_id: { type: string, example: "111111111111" }
        alias: { type: string, example: "prod" }
        environment: { type: string, example: "production" }
        regions: { type: array, items: { type: string }, example: ["us-east-1", "us-west-2"] }
        access_mode: { type: string, enum: [assume_role, simulated] }
        role_arns: { type: object, additionalProperties: { type: string } }
        external_id: { type: string }
        session_prefix: { type: string }
        state: { type: string, enum: [pending, connected, degraded, suspended, error] }
        state_reason: { type: string }
        granted_scopes: { type: array, items: { type: string, enum: [read, analyze, plan, execute] } }
        missing_actions: { type: array, items: { type: string } }
        is_payer: { type: boolean }
        cur_bucket: { type: string }
        cur_prefix: { type: string }
        connected_at: { type: string, format: date-time }
        last_verified_at: { type: string, format: date-time }
        last_discovery_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    RegisterAccountRequest:
      type: object
      required: [account_id]
      properties:
        account_id: { type: string, example: "111111111111" }
        alias: { type: string }
        environment: { type: string, example: "production" }
        regions: { type: array, items: { type: string } }
        access_mode: { type: string, enum: [assume_role, simulated], default: assume_role }
        role_arns:
          type: object
          additionalProperties: { type: string }
          description: "Map of role scope (read|analyze|plan|execute) to the ARN CloudOptix should assume for it."
        is_payer: { type: boolean }
        cur_bucket: { type: string }
        cur_prefix: { type: string }
    SuspendAccountRequest:
      type: object
      properties:
        reason: { type: string }

    # -------------------------------------------------------------
    # Discovery
    # -------------------------------------------------------------
...
    CostPoint:
      type: object
      properties:
        period: { $ref: "#/components/schemas/Period" }
        amount: { $ref: "#/components/schemas/Money" }
    CostSeries:
      type: object
      properties:
        granularity: { type: string, enum: [hourly, daily, monthly] }
        points: { type: array, items: { $ref: "#/components/schemas/CostPoint" } }
        currency: { type: string }
    BreakdownItem:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        amount: { $ref: "#/components/schemas/Money" }
        share: { type: number }
        prior_amount: { $ref: "#/components/schemas/Money" }
        change_pct: { type: number }
        resource_count: { type: integer }
    CostBreakdown:
      type: object
      properties:
        dimension: { type: string }
        period: { $ref: "#/components/schemas/Period" }
        total: { $ref: "#/components/schemas/Money" }
        items: { type: array, items: { $ref: "#/components/schemas/BreakdownItem" } }
    CostForecast:
      type: object
      properties:
        period: { $ref: "#/components/schemas/Period" }
        expected: { $ref: "#/components/schemas/Money" }
        low: { $ref: "#/components/schemas/Money" }
        high: { $ref: "#/components/schemas/Money" }
        method: { type: string, example: "holt_winters" }
        confidence: { $ref: "#/components/schemas/Confidence" }
        based_on_days: { type: integer }
        note: { type: string }
...
    CostSummary:
      type: object
      properties:
        period: { $ref: "#/components/schemas/Period" }
        total: { $ref: "#/components/schemas/Money" }
        daily_average: { $ref: "#/components/schemas/Money" }
        month_to_date: { $ref: "#/components/schemas/Money" }
        prior_month: { $ref: "#/components/schemas/Money" }
        change_pct: { type: number }
        forecast: { $ref: "#/components/schemas/CostForecast" }
        by_service: { $ref: "#/components/schemas/CostBreakdown" }
        by_account: { $ref: "#/components/schemas/CostBreakdown" }
        by_environment: { $ref: "#/components/schemas/CostBreakdown" }
        by_application: { $ref: "#/components/schemas/CostBreakdown" }
        trend: { $ref: "#/components/schemas/CostSeries" }
        open_anomalies: { type: integer }
        last_ingested_at: { type: string, format: date-time }
        freshness: { type: string }

    # -------------------------------------------------------------
    # Economics / Cost SLOs
    # -------------------------------------------------------------
...
    RecommendationSummary:
      type: object
      properties:
        open: { type: integer }
        total_monthly_saving: { $ref: "#/components/schemas/Money" }
        by_category: { type: object, additionalProperties: { type: integer } }
        saving_by_category: { type: object, additionalProperties: { $ref: "#/components/schemas/Money" } }
        by_risk: { type: object, additionalProperties: { type: integer } }
        auto_executable: { type: integer }
        awaiting_approval: { type: integer }
        mutually_exclusive_alternatives:
          type: integer
          description: >-
            How many open recommendations are alternatives to another one and
            therefore excluded from total_monthly_saving and
            saving_by_category. Counts include them; money does not.
...
    ApprovalContext:
      type: object
      properties:
        monthly_saving: { $ref: "#/components/schemas/Money" }
        annual_saving: { $ref: "#/components/schemas/Money" }
        monthly_cost_delta: { $ref: "#/components/schemas/Money" }
        confidence: { $ref: "#/components/schemas/Confidence" }
        risk_level: { $ref: "#/components/schemas/RiskLevel" }
        blast_summary: { type: string }
        environment: { type: string }
        affected_resources: { type: array, items: { type: string } }
        rollback_plan: { type: string }
        validation_plan: { type: string }
        policy_reason: { type: string }
        diff: { type: string }
    ApprovalResponse:
      type: object
      properties:
        principal: { type: string }
        role: { type: string }
        approved: { type: boolean }
        comment: { type: string }
        at: { type: string, format: date-time }
        ip_address: { type: string }
        user_agent: { type: string }
    ApprovalRequest:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        subject_kind: { type: string, enum: [execution_plan, policy, aws_connection, commitment_purchase] }
        subject_id: { type: string }
        title: { type: string }
        summary: { type: string }
        context: { $ref: "#/components/schemas/ApprovalContext" }
        policy_decision_id: { type: string }
        required_roles: { type: array, items: { type: string } }
        min_approvals: { type: integer }
        require_distinct_approver: { type: boolean }
        state: { type: string, enum: [pending, approved, rejected, expired, withdrawn] }
        responses: { type: array, items: { $ref: "#/components/schemas/ApprovalResponse" } }
        requested_by: { type: string }
        requested_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        decided_at: { type: string, format: date-time }
        execute_after: { type: string, format: date-time }
    ApprovalPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/ApprovalRequest" } }
...
    AutonomousRunResult:
      type: object
      properties:
        considered: { type: integer }
        planned: { type: integer }
        executed: { type: integer }
        skipped: { type: integer }
        failed: { type: integer }
        rolled_back: { type: integer }
        monthly_saving: { $ref: "#/components/schemas/Money" }
        skip_reasons: { type: object, additionalProperties: { type: integer } }
        duration_ms: { type: integer }
...